chore(hooks): [useLocale] autocomplete locale field keys (#21876)
